fix(replies): fix replies from Status and to Status messages
Needed for https://github.com/status-im/status-desktop/issues/16323

There were multiple problems.

1. The ParentID was not set correctly when a message was sent from Status
2. We didn't send back the ID of messages that we sent, so the gateway couldn't associate the new messages to the old ones
3. Missing param in the toml config `PreserveThreading`
4. Discord messages are weirdly formatted when a reply is send. It cannot use webhooks, so we need to pass the username directly in the message (it's ugly, but it works)

Finally, there was an issue on the status-go side. We tried to use the discord message ID to find the replies, but the bridge already converts it to the right ID, so we could just use it directly